@@ -28,14 +29,17 @@
   "license": "MIT",
   "author": "Paul Souche <psouche@santech.fr>",
   "peerDependencies": {
-    "@angular/core": "^5.0.0",
-    "@ionic-native/camera": "^4.0.0",
-    "@ionic-native/core": "^4.0.0",
-    "@ionic-native/file": "^4.0.0",
-    "@ionic-native/keyboard": "^4.0.0",
-    "@ionic-native/network": "^4.0.0",
-    "@ionic-native/onesignal": "^4.0.0",
-    "@ionic-native/splash-screen": "^4.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@ionic-native/camera": "^5.0.0-beta.0",
+    "@ionic-native/core": "^5.0.0-beta.0",
+    "@ionic-native/file": "^5.0.0-beta.0",
+    "@ionic-native/keyboard": "^5.0.0-beta.0",
+    "@ionic-native/network": "^5.0.0-beta.0",
+    "@ionic-native/onesignal": "^5.0.0-beta.0",
+    "@ionic-native/splash-screen": "^5.0.0-beta.0",
+    "@ionic/angular": "^4.0.0-beta.0",
+    "@ionic/core": "^4.0.0-beta.0",
     "@ionic/storage": "^2.0.0",
     "@santech/analytics-core": "^7.0.0-alpha.0",
     "@santech/analytics-integration": "^7.0.0-alpha.0",
