@@ -8,10 +8,11 @@
   "types": "dist/index.d.ts",
   "typings": "dist/santech-angular-pica.d.ts",
   "scripts": {
-    "build": "npm run clean && ng-packagr -p ng-package.json && npm run docs",
-    "clean": "rimraf dist docs coverage dist.tgz",
+    "build": "npm run clean && ts-node build.ts && npm run docs && npm run package",
+    "clean": "rimraf dist docs coverage .ng-build dist.tgz",
     "docs": "typedoc src/index.ts",
-    "lint": "tslint \"**/*.ts\" \"**/*.js\"",
+    "lint": "tslint \"**/*.ts\" \"**/*.js\" -e \"node_modules/**\"",
+    "package": "npm pack dist/ && mv santech-angular-pica-*.tgz dist.tgz",
     "test": "jest --coverage",
     "test-watch": "jest --watch"
   },
@@ -28,38 +29,39 @@
   "author": "Paul Souche <psouche@santech.fr>",
   "license": "MIT",
   "peerDependencies": {
-    "@angular/core": "^5.0.0",
+    "@angular/core": "^6.0.0",
     "@santech/angular-common": "^7.0.0-alpha.12",
     "@santech/angular-platform": "^7.0.0-alpha.12",
     "exif-js": "^2.0.0",
     "pica": "^4.0.0",
-    "rxjs": "^5.0.0"
+    "rxjs": "^6.0.0"
   },
   "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/common": "*",
+    "@angular/compiler": "*",
+    "@angular/compiler-cli": "*",
+    "@angular/core": "*",
+    "@angular/platform-browser": "*",
+    "@angular/platform-browser-dynamic": "*",
     "@santech/angular-common": "^7.0.0-alpha.14",
     "@santech/angular-platform": "^7.0.0-alpha.14",
     "@types/jest": "*",
     "@types/node": "*",
     "jest": "*",
     "jest-localstorage-mock": "*",
-    "jest-preset-angular": "*",
-    "ng-packagr": "*",
+    "jest-preset-angular": "^6.0.0-rc.0",
+    "ng-packagr": "3.0.6",
     "rimraf": "*",
-    "rxjs": "^5.0.0",
+    "rxjs": "*",
     "tsickle": "*",
     "tslib": "*",
     "typedoc": "*",
-    "typescript": "*",
+    "typescript": "^2.0.0",
     "whatwg-fetch": "*",
     "zone.js": "*"
   },
   "jest": {
+    "testURL": "http://localhost",
     "transform": {
       "^.+\\.(ts|js|html)$": "<rootDir>../../node_modules/jest-preset-angular/preprocessor.js"
     },
