@@ -24,35 +24,37 @@
   },
   "license": "MIT",
   "resolutions": {
-    "@segment/analytics.js-core/**/uuid": "3.2.1",
-    "lolex": "2.3.2"
+    "@segment/analytics.js-core/**/uuid": "3.3.2"
   },
   "dependencies": {
-    "@angular/core": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ionic-native/camera": "^4.0.0",
-    "@ionic-native/core": "^4.0.0",
-    "@ionic-native/file": "^4.0.0",
-    "@ionic-native/keyboard": "^4.0.0",
-    "@ionic-native/network": "^4.0.0",
-    "@ionic-native/onesignal": "^4.0.0",
-    "@ionic-native/splash-screen": "^4.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "@ionic-native/camera": "^5.0.0-beta.0",
+    "@ionic-native/core": "^5.0.0-beta.0",
+    "@ionic-native/file": "^5.0.0-beta.0",
+    "@ionic-native/keyboard": "^5.0.0-beta.0",
+    "@ionic-native/network": "^5.0.0-beta.0",
+    "@ionic-native/onesignal": "^5.0.0-beta.0",
+    "@ionic-native/splash-screen": "^5.0.0-beta.0",
+    "@ionic/angular": "^4.0.0-beta.0",
+    "@ionic/core": "^4.0.0-beta.0",
     "@ionic/storage": "^2.0.0",
-    "@ngx-translate/core": "^9.0.0",
+    "@ngx-translate/core": "^10.0.0",
     "@stomp/stompjs": "^4.0.0",
+    "babel-core": "^6.0.0",
     "exif-js": "^2.0.0",
-    "ionic-angular": "^3.0.0",
     "ionic-cache": "^3.0.0",
     "pica": "^4.0.0",
-    "rxjs": "^5.0.0",
+    "rxjs": "^6.0.0",
+    "rxjs-compat": "^6.0.0",
     "sockjs-client": "^1.0.0",
     "tslib": "^1.0.0",
     "zone.js": "^0.8.0"
   },
   "devDependencies": {
-    "lerna": "^2.0.0"
+    "lerna": "^3.0.0-rc.0"
   }
 }
