@@ -36,20 +36,20 @@ describe('Santech analytics Integration', () => {
             version: 'version',
           },
           page: {
-            path: 'blank',
+            path: '/',
             referrer: '',
             search: '',
             title: '',
-            url: 'about:blank',
+            url: 'http://localhost/',
           },
         },
         integrations: {},
         properties: {
-          path: 'blank',
+          path: '/',
           referrer: '',
           search: '',
           title: '',
-          url: 'about:blank',
+          url: 'http://localhost/',
         },
         timestamp: jasmine.any(Date),
         type: 'page',
