@@ -1,3 +1,3 @@
-import { AlertOptions } from 'ionic-angular';
+import { AlertOptions } from '@ionic/core';
 
 export type confirmFunction = (confirmHandler: () => void, cancelHandler: () => void) => AlertOptions;
