@@ -1,10 +1,10 @@
-import { CacheService, CameraService, OneSignalService } from '@santech/angular-ionic';
+import { CameraService, OneSignalService, RepositoryService } from '@santech/angular-ionic';
 import { createJestSpyObj, filterPrivate, SantechSpyObject } from '@santech/core/testing';
 
 export * from './ionic';
 
 export const cacheServiceMethods = Object
-  .getOwnPropertyNames(CacheService.prototype).filter(filterPrivate);
+  .getOwnPropertyNames(RepositoryService.prototype).filter(filterPrivate);
 
 export const cameraServiceMethods = Object
   .getOwnPropertyNames(CameraService.prototype).filter(filterPrivate);
@@ -12,22 +12,22 @@ export const cameraServiceMethods = Object
 export const oneSignalServiceMethods = Object
   .getOwnPropertyNames(OneSignalService.prototype).filter(filterPrivate);
 
-let spyCacheService: SantechSpyObject<CacheService>;
 let spyCameraService: SantechSpyObject<CameraService>;
 let spyOneSignalService: SantechSpyObject<OneSignalService>;
+let spyRepositoryService: SantechSpyObject<RepositoryService>;
 
 if (typeof jasmine !== 'undefined' && typeof jasmine.createSpyObj === 'function') {
-  spyCacheService = jasmine.createSpyObj('spyCacheService', cacheServiceMethods);
   spyCameraService = jasmine.createSpyObj('spyCameraService', cameraServiceMethods);
   spyOneSignalService = jasmine.createSpyObj('spyOneSignalService', oneSignalServiceMethods);
+  spyRepositoryService = jasmine.createSpyObj('spyCacheService', cacheServiceMethods);
 } else if (typeof jest !== 'undefined') {
-  spyCacheService = createJestSpyObj(cacheServiceMethods);
   spyCameraService = createJestSpyObj(cameraServiceMethods);
   spyOneSignalService = createJestSpyObj(oneSignalServiceMethods);
+  spyRepositoryService = createJestSpyObj(cacheServiceMethods);
 }
 
 export {
-  spyCacheService,
   spyCameraService,
   spyOneSignalService,
+  spyRepositoryService,
 };
